Add files via upload
Today, computers are widely used in various fields. Efficiency of the computer is very important for better functioning. In computing, optimization is the process of modifying a system to make some features of it work more efficiently or use fewer resources. For instance, a computer program may be optimized so that it runs faster, or to run with less memory requirements or other resources, or to consume less energy. Our project is an optimisation model which covers aspects like power efficiency of the CPU fans, consolidating data on minimum number of hard drives and designing a cost-efficient computer network. It is an effort to help the small-scale companies and local people to assemble their computers and use it effectively and face lesser problems in the long run. This project shall be open source to broaden its reach and serve as a useful resource to the society.